Ajustes finales en variaciones y especificaciones

+++ api_ecommerce/app/Http/Controllers/Admin/Product/AttributeProductController.php
@@ -89,14 +89,6 @@ class AttributeProductController extends Controller
             ],
         ]);
     }
-    public function destroy_propertie($id) {
-        $propertie = Propertie::findOrFail($id);
-        $propertie->delete();//IMPORTANTE VALIDACION
-        return response()->json([
-            'message' => 200,
-        ]);
-
-    }
 
     /**
      * Display the specified resource.
@@ -143,9 +135,28 @@ class AttributeProductController extends Controller
     public function destroy(string $id)
     {
         $attribute = Attribute::findOrFail($id);
+
+        if($attribute->specifications()->count() > 0 || 
+        $attribute->variations()->count() > 0){
+            return response()->json(['message' => 403,"message_text" => "EL ATRIBUTO YA ESTA RELACIONADO CON ALGUNOS O UN PRODUCTO"]);
+        }
         $attribute->delete();//IMPORTANTE VALIDACION
         return response()->json([
             'message' => 200,
         ]);
+    }
+
+    
+    public function destroy_propertie($id) {
+        $propertie = Propertie::findOrFail($id);
+        $propertie->delete();//IMPORTANTE VALIDACION
+        if($propertie->specifications()->count() > 0 || 
+        $propertie->variations()->count() > 0){
+            return response()->json(['message' => 403,"message_text" => "LA PROPIEDAD YA ESTA RELACIONADO CON ALGUNOS O UN PRODUCTO"]);
+        }
+        return response()->json([
+            'message' => 200,
+        ]);
+
     }
 }
